@@ -1,7 +1,7 @@
 # Created: 11.03.2011
 # Copyright (c) 2011-2018, Manfred Moitzi
 # License: MIT License
-from __future__ import unicode_literals
+
 from datetime import datetime
 import io
 import logging
