@@ -1,12 +1,9 @@
 # Created: 08.04.2018
 # Copyright (c) 2018, Manfred Moitzi
 # License: MIT-License
-from __future__ import unicode_literals
-
 from ezdxf.lldxf.const import DXFValueError
 from ezdxf.lldxf import loader
 from ezdxf.lldxf.packedtags import TagList
-from ezdxf.tools.c23 import isstring
 
 from .dxfobjects import ExtendedTags, DefSubclass, DXFAttributes
 from .dxfobjects import none_subclass, DXFObject
@@ -59,7 +56,7 @@ class LayerFilter(DXFObject):
 
     @layers.setter
     def layers(self, names):
-        if isstring(names):
+        if isinstance(names, str):
             raise DXFValueError('requires iterable but not string')
         self.layers[:] = list(names)
 
