@@ -2,7 +2,7 @@
 # Created: 21.03.2011
 # Copyright (c) 2011-2018, Manfred Moitzi
 # License: MIT License
-from __future__ import unicode_literals
+
 from math import floor
 from datetime import datetime
 
